# frontend/docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Sagara Mission Control API
  version: 1.0.0
  description: |
    Canonical Sagara Mission Control API Specification (Contract V1).
    Frozen for future FastAPI backend implementation.
    Adheres strictly to snake_case naming, ISO 8601 UTC timestamps, opaque string IDs,
    pessimistic mutation locking, explicit idempotency keys, and UNKNOWN != ZERO semantics.
servers:
  - url: http://localhost:8000/api/v1
    description: Local Mission Control Backend
  - url: https://mission-control.sagara.local/api/v1
    description: Production Internal Gateway
paths:
...
  /tasks:
    get:
      summary: List Tasks
      operationId: listTasks
      parameters:
        - name: state
          in: query
          schema:
            type: string
        - name: priority
          in: query
          schema:
            type: string
        - name: agent_id
          in: query
          schema:
            type: string
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Task projections
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaskDto'
    post:
      summary: Create Task
      operationId: createTask
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskDto'
      responses:
        '201':
          description: Task created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /tasks/{task_id}:
    get:
      summary: Get Task by ID
      operationId: getTaskById
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Task details
          headers:
            ETag:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskDto'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      summary: Update Task
      operationId: updateTask
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: string
        - name: If-Match
          in: header
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskDto'
      responses:
        '200':
          description: Task updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskDto'
        '409':
          description: Concurrency version conflict (ETag mismatch)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
  /tasks/{task_id}/dispatch:
    post:
      summary: Dispatch Task for Execution
      operationId: dispatchTask
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: string
        - name: Idempotency-Key
          in: header
          required: false
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Task transitioned to DISPATCHING/QUEUED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskDto'
        '400':
          description: Invalid state for dispatch
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
        '409':
          description: Task already dispatched or locked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
  /tasks/{task_id}/cancel:
    post:
      summary: Cancel Task
      operationId: cancelTask
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: string
        - name: Idempotency-Key
          in: header
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Task marked as CANCELLED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskDto'
        '409':
          $ref: '#/components/responses/Conflict'
...
components:
  schemas:
    ApiErrorDto:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              example: APPROVAL_ALREADY_RESOLVED
            message:
              type: string
              example: This approval gate has already been decided.
            details:
              type: object
              nullable: true
            correlation_id:
              type: string
              nullable: true
              example: corr-20260909-001
...
    TaskDto:
      type: object
      required:
        - id
        - title
        - state
        - priority
        - created_at
      properties:
        id:
          type: string
          example: tsk-01
        title:
          type: string
        description:
          type: string
          nullable: true
        state:
          type: string
          enum:
            - DRAFT
            - READY
            - QUEUED
            - DISPATCHING
            - RUNNING
            - AWAITING_APPROVAL
            - BLOCKED
            - COMPLETED
            - FAILED
            - CANCELLED
        priority:
          type: string
          enum: [LOW, MEDIUM, HIGH, CRITICAL]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
          nullable: true
        assigned_agent_id:
          type: string
          nullable: true
        requested_skills:
          type: array
          items:
            type: string
          nullable: true
        progress:
          type: object
          nullable: true
          properties:
            completed:
              type: integer
              nullable: true
            total:
              type: integer
              nullable: true
            label:
              type: string
              nullable: true
    CreateTaskDto:
      type: object
      required:
        - title
        - priority
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        priority:
          type: string
          enum: [LOW, MEDIUM, HIGH, CRITICAL]
        state:
          type: string
          enum: [DRAFT, READY, QUEUED]
        assigned_agent_id:
          type: string
          nullable: true
        requested_skills:
          type: array
          items:
            type: string
          nullable: true
    UpdateTaskDto:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        priority:
          type: string
        state:
          type: string
        assigned_agent_id:
          type: string
...
  responses:
    BadRequest:
      description: Malformed syntax or invalid request format
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorDto'
    NotFound:
      description: Resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorDto'
    Conflict:
      description: State conflict or concurrent modification
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorDto'
...